+++ packages/envied/lib/src/envied_base.dart
@@ -86,6 +86,10 @@ final class Envied {
   /// Escapes single quotes and newlines in the value.
   final bool rawStrings;
 
+  /// A seed can be provided if the obfuscation randomness needs to remain
+  /// reproducible across builds.
+  final int? randomSeed;
+
   const Envied({
     String? path,
     bool? requireEnvFile,
@@ -95,6 +99,7 @@ final class Envied {
     this.useConstantCase = false,
     this.interpolate = true,
     this.rawStrings = false,
+    this.randomSeed,
   })  : path = path ?? '.env',
         requireEnvFile = requireEnvFile ?? false;
 }
@@ -160,6 +165,10 @@ final class EnviedField {
   /// Escapes single quotes and newlines in the value.
   final bool? rawString;
 
+  /// A seed can be provided if the obfuscation randomness needs to remain
+  /// reproducible across builds.
+  final int? randomSeed;
+
   const EnviedField({
     this.varName,
     this.obfuscate,
@@ -168,5 +177,6 @@ final class EnviedField {
     this.useConstantCase,
     this.interpolate,
     this.rawString,
+    this.randomSeed,
   });
 }